% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is_symmetric.R
\name{is.symmetric}
\alias{is.symmetric}
\title{Tests whether the given matrix is symmetric.}
\usage{
is.symmetric(mat)
}
\arguments{
\item{mat}{Matrix to be tested for symmetry.}
}
\value{
Whether the matrix is symmetric.
}
\description{
Tests whether the given matrix is symmetric.
}

