% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pekalska.R
\name{pekalska}
\alias{pekalska}
\title{Pekalska's approach to speeding up Sammon's mapping.}
\usage{
pekalska(D, sample.indices = NULL, Ys = NULL)
}
\arguments{
\item{D}{dist object or distances matrix.}

\item{sample.indices}{The indices of subsamples.}

\item{Ys}{The subsample mapping (k-dimensional).}
}
\value{
The low-dimensional representation of the data.
}
\description{
Creates a k-dimensional representation of the data. As input, a subsample and
its k-dimensional mapping are required. The method approximates the subsample
mapping to a linear mapping based on the distances matrix of the subsample
and then applies the same mapping to all instances.
}
\references{
Pekalska, E., de Ridder, D., Duin, R. P., & Kraaijveld, M. A.
(1999). A new method of generalizing Sammon mapping with application to
algorithm speed-up (pp. 221-228).
}

