% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lamp}
\alias{lamp}
\title{Local Affine Multidimensional Projection}
\usage{
lamp(X, sample.indices = NULL, Ys = NULL)
}
\arguments{
\item{X}{A data frame or matrix to be projected.}

\item{sample.indices}{The indices of data points in X considered as samples.}

\item{Ys}{Initial 2D configuration of the data samples (will be ignored if
sample.indices is NULL).}
}
\value{
The 2D representation of the data.
}
\description{
Creates a 2D representation of the data based on an initial projectec sample.
}
\examples{
# Iris example
proj = lamp(iris[, 1:4])
plot(proj, col = iris$Species)
}
\references{
Joia, P.; Paulovich, F.V.; Coimbra, D.; Cuminato, J.A.; Nonato,
  L.G., "Local Affine Multidimensional Projection," Visualization and
  Computer Graphics, IEEE Transactions on , vol.17, no.12, pp.2563,2571,
  Dec. 2011
}

